MKB-659 rename Locality to Branch
